# openapi.json
{
  "components": {
    "schemas": {
...
      "DomainsResponse": {
        "description": "All configured domains plus their physical presence.",
        "examples": [
          {
            "domains": [
              {
                "configured": true,
                "domain": "hr",
                "exists": true
              }
            ],
            "store": "chroma",
            "tenant_mode": "off"
          }
        ],
        "properties": {
          "domains": {
            "items": {
              "$ref": "#/components/schemas/DomainInfo"
            },
            "title": "Domains",
            "type": "array"
          },
          "store": {
            "title": "Store",
            "type": "string"
          },
          "tenant_mode": {
            "title": "Tenant Mode",
            "type": "string"
          }
        },
        "required": [
          "store",
          "tenant_mode",
          "domains"
        ],
        "title": "DomainsResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Consistent error envelope used by every non-2xx response.",
        "examples": [
          {
            "error": {
              "code": "forbidden",
              "details": {
                "required_scope": "kb:hr:write"
              },
              "message": "Insufficient scope for domain 'hr'",
              "request_id": "3f1c1a0e9f4b4c0a8f2a"
            }
          }
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JobResponse": {
        "description": "Status of a background ingestion job.",
        "properties": {
          "job": {
            "$ref": "#/components/schemas/Job"
          }
        },
        "required": [
          "job"
        ],
        "title": "JobResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerJWT": {
        "bearerFormat": "JWT",
        "description": "JWT signed with the shared secret (`JWT_SECRET_KEY`, HS256 by default). Mint a development token with `python scripts/mint_token.py`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "platform@example.com",
      "name": "Platform Engineering"
    },
    "description": "Retrieval-augmented agent service for the NexusOne ERP.\n\n**Domain separated knowledge bases.** Each business domain (`hr`, `finance`,\n`engineering`, `it`, `company`, ...) is an isolated index. A query never crosses\ndomains unless the caller asks for it, and access is granted per domain through\nJWT scopes of the form `kb:<domain>:read` / `kb:<domain>:write`.\n\n**Pluggable everything.** The vector store (`chroma` | `faiss` | `pinecone`),\nthe chat model (`openai` | `anthropic`) and the embedding model\n(`openai` | `huggingface`) are selected purely through environment variables.\n\n**Authentication.** Every endpoint except `/health`, `/ready` and the docs\nrequires `Authorization: Bearer <jwt>`; tokens are signed by the calling backend\nwith a shared HMAC secret. Click **Authorize** to try the endpoints below.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "NexusOne Agent Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/kb/domains": {
      "get": {
        "description": "Returns every domain configured through `KB_DOMAINS` together with whether a physical collection/namespace currently exists for it. Requires `kb:admin`.",
        "operationId": "list_domains_kb_domains_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token lacks the required scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown domain or invalid request body."
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "summary": "List knowledge base domains",
        "tags": [
          "kb"
        ]
      }
    },
    "/kb/jobs/{job_id}": {
      "get": {
        "description": "Poll the status of an ingestion job returned by an `mode=async` ingest call.",
        "operationId": "get_job_kb_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token lacks the required scope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown job."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown domain or invalid request body."
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "summary": "Get a background ingestion job",
        "tags": [
          "kb"
        ]
      }
    },
    "/kb/{domain}": {
      "delete": {
        "description": "Deletes the whole collection/namespace for a domain. Requires `kb:admin`.",
        "operationId": "drop_domain_kb__domain__delete",
        "parameters": [
          {
            "description": "Business domain, must be one of KB_DOMAINS.",
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "description": "Business domain, must be one of KB_DOMAINS.",
              "examples": [
                "hr",
                "finance"
              ],
              "title": "Domain",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token lacks the required scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown domain or invalid request body."
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "summary": "Drop an entire domain",
        "tags": [
          "kb"
        ]
      }
    },
    "/kb/{domain}/documents": {
      "delete": {
        "description": "Removes chunks by explicit id or by metadata filter (for example every chunk of one `doc_id`). Requires scope `kb:<domain>:write`.",
        "operationId": "delete_documents_kb__domain__documents_delete",
        "parameters": [
          {
            "description": "Business domain, must be one of KB_DOMAINS.",
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "description": "Business domain, must be one of KB_DOMAINS.",
              "examples": [
                "hr",
                "finance"
              ],
              "title": "Domain",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDocumentsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token lacks the required scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown domain or invalid request body."
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "summary": "Delete chunks from a domain",
        "tags": [
          "kb"
        ]
      }
    },
    "/kb/{domain}/ingest/files": {
      "post": {
        "description": "Accepts a multipart upload with one or more files (`pdf`, `docx`, `txt`, `md`, `csv`, `json`, `html`). Files are chunked, embedded and de-duplicated by content hash. Small batches run inline; anything above `BACKGROUND_INGEST_THRESHOLD` returns a job id to poll at `GET /kb/jobs/{id}`.\n\nRequires scope `kb:<domain>:write`.",
        "operationId": "ingest_files_kb__domain__ingest_files_post",
        "parameters": [
          {
            "description": "Business domain, must be one of KB_DOMAINS.",
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "description": "Business domain, must be one of KB_DOMAINS.",
              "examples": [
                "hr",
                "finance"
              ],
              "title": "Domain",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_files_kb__domain__ingest_files_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token lacks the required scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown domain or invalid request body."
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "summary": "Ingest uploaded files into a domain",
        "tags": [
          "kb"
        ]
      }
    },
    "/kb/{domain}/ingest/text": {
      "post": {
        "description": "Chunks and embeds a string. Useful for ERP records, generated summaries or policy snippets. Requires scope `kb:<domain>:write`.",
        "operationId": "ingest_text_kb__domain__ingest_text_post",
        "parameters": [
          {
            "description": "Business domain, must be one of KB_DOMAINS.",
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "description": "Business domain, must be one of KB_DOMAINS.",
              "examples": [
                "hr",
                "finance"
              ],
              "title": "Domain",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token lacks the required scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown domain or invalid request body."
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "summary": "Ingest raw text into a domain",
        "tags": [
          "kb"
        ]
      }
    },
    "/kb/{domain}/ingest/urls": {
      "post": {
        "description": "Fetches each URL, extracts readable text and ingests it. URLs resolving to private, loopback or link-local addresses are rejected (SSRF guard). Requires scope `kb:<domain>:write`.",
        "operationId": "ingest_urls_kb__domain__ingest_urls_post",
        "parameters": [
          {
            "description": "Business domain, must be one of KB_DOMAINS.",
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "description": "Business domain, must be one of KB_DOMAINS.",
              "examples": [
                "hr",
                "finance"
              ],
              "title": "Domain",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestUrlsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token lacks the required scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown domain or invalid request body."
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "summary": "Ingest web pages into a domain",
        "tags": [
          "kb"
        ]
      }
    },
    "/kb/{domain}/reindex": {
      "post": {
        "description": "Reads every chunk, re-embeds it with the currently configured embedding model into a staging collection, then swaps the staging collection over the live one. Use after changing `EMBEDDING_MODEL`. Requires `kb:admin`.",
        "operationId": "reindex_kb__domain__reindex_post",
        "parameters": [
          {
            "description": "Business domain, must be one of KB_DOMAINS.",
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "description": "Business domain, must be one of KB_DOMAINS.",
              "examples": [
                "hr",
                "finance"
              ],
              "title": "Domain",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReindexResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token lacks the required scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown domain or invalid request body."
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "summary": "Re-embed a domain and swap it in atomically",
        "tags": [
          "kb"
        ]
      }
    },
...
    "/kb/{domain}/stats": {
      "get": {
        "description": "Store type, physical collection/namespace name, vector count, embedding model and dimension. Requires `kb:admin`.",
        "operationId": "domain_stats_kb__domain__stats_get",
        "parameters": [
          {
            "description": "Business domain, must be one of KB_DOMAINS.",
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "description": "Business domain, must be one of KB_DOMAINS.",
              "examples": [
                "hr",
                "finance"
              ],
              "title": "Domain",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainStats"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token lacks the required scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown domain or invalid request body."
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "summary": "Domain statistics",
        "tags": [
          "kb"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This deployment",
      "url": "/"
    }
  ],
...
}